Fix issues with Generate Type tests
