+++ src/EditorFeatures/TestUtilities/Diagnostics/GenerateType/GenerateTypeTestState.cs
@@ -8,6 +8,8 @@ using System.Linq;
 using Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces;
 using Microsoft.CodeAnalysis.GenerateType;
 using Microsoft.CodeAnalysis.ProjectManagement;
+using Roslyn.Utilities;
+using Xunit;
 
 namespace Microsoft.CodeAnalysis.Editor.UnitTests.Diagnostics.GenerateType
 {
@@ -22,31 +24,15 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Diagnostics.GenerateType
         public Project TriggeredProject { get; }
         public string TypeName { get; }
 
-        public static GenerateTypeTestState Create(
-            string initial,
+        public GenerateTypeTestState(
+            TestWorkspace workspace,
             string projectToBeModified,
             string typeName,
-            string existingFileName,
-            string languageName)
+            string existingFileName)
         {
-            var workspace = TestWorkspace.IsWorkspaceElement(initial)
-                ? TestWorkspace.Create(initial)
-                : languageName == LanguageNames.CSharp
-                  ? TestWorkspace.CreateCSharp(initial)
-                  : TestWorkspace.CreateVisualBasic(initial);
-
-            return new GenerateTypeTestState(projectToBeModified, typeName, existingFileName, workspace);
-        }
-
-        private GenerateTypeTestState(string projectToBeModified, string typeName, string existingFileName, TestWorkspace testWorkspace)
-        {
-            Workspace = testWorkspace;
+            Workspace = workspace;
             _testDocument = Workspace.Documents.SingleOrDefault(d => d.CursorPosition.HasValue);
-
-            if (_testDocument == null)
-            {
-                throw new ArgumentException("markup does not contain a cursor position", "workspace");
-            }
+            Contract.ThrowIfNull(_testDocument, "markup does not contain a cursor position");
 
             TriggeredProject = Workspace.CurrentSolution.GetProject(_testDocument.Project.Id);
 
@@ -58,10 +44,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Diagnostics.GenerateType
             else
             {
                 ProjectToBeModified = Workspace.CurrentSolution.Projects.FirstOrDefault(proj => proj.Name.Equals(projectToBeModified));
-                if (ProjectToBeModified == null)
-                {
-                    throw new ArgumentException("Project with the given name does not exist", "workspace");
-                }
+                Contract.ThrowIfNull(ProjectToBeModified, "Project with the given name does not exist");
             }
 
             InvocationDocument = Workspace.CurrentSolution.GetDocument(_testDocument.Id);
